requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq

sentence-transformers
faiss-cpu
pypdf
